Put tests failing on unprivileged container behind a CONTAINERTEST env-var flag.

+++ tests/p_dovecot/dovecot_imap_login.sh
@@ -4,6 +4,11 @@
 
 t_Log "Running $0 - adding imaptest local user account + attempting IMAP login"
 
+if [ "$CONTAINERTEST" -eq "1" ]; then
+    t_Log "Running in container -> SKIP"
+    exit 0
+fi
+
 { userdel -rf imaptest; useradd imaptest && echo imaptest | passwd --stdin imaptest; } &>/dev/null
 
 # creating maildir in HOME, else test will fail at first try
