+++ tests/p_vsftpd/vsftpd_localusers_login.sh
@@ -3,6 +3,11 @@
 
 t_Log "Running $0 - vsFTPd local user can login test."
 
+if [ "$CONTAINERTEST" -eq "1" ]; then
+    t_Log "Running in container -> SKIP"
+    exit 0
+fi
+
 # Add a test ftp user
 { userdel -rf ftptest; useradd ftptest && echo ftptest | passwd --stdin ftptest; } &>/dev/null
